+++ packages/plugin/finder/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @umac-js/finder
 
+## 2.2.4
+
+### Patch Changes
+
+- fix utils and core errors
+
+- Updated dependencies []:
+  - @umac-js/utils@2.2.4
+
 ## 2.2.3
 
 ### Patch Changes