+++ packages/repo-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @umac-js/repo-config
 
+## 2.2.4
+
+### Patch Changes
+
+- fix utils and core errors
+
 ## 2.2.3
 
 ### Patch Changes